requirements.txt
flask==2.0.1
openai==0.10.1
langchain==0.0.98
faiss-cpu==1.7.2
python-dotenv==0.19.2
